Refactor MinigameSolver and MinigameHandler
