+++ Neuro/Minigames/MinigameSolver.cs
@@ -1,47 +1,12 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using Il2CppInterop.Runtime;
-using Neuro.Cursor;
+﻿using System.Collections;
 using UnityEngine;
 
-namespace Neuro.Minigames.Completion;
+namespace Neuro.Minigames;
 
 public abstract class MinigameSolver
 {
     // The larger this constant is, the longer the delays are when solving minigames.
     public const float DELAY_MULTIPLIER = 1;
-
-    static MinigameSolver()
-    {
-        MinigameSolvers = Assembly.GetExecutingAssembly().GetTypes()
-            .Where(t => t.GetCustomAttribute<MinigameSolverAttribute>() is { })
-            .Where(t => t.IsAssignableTo(typeof(MinigameSolver)))
-            .Select(solverType => (solverType.GetCustomAttribute<MinigameSolverAttribute>()!.Types, solverType))
-            .SelectMany(t => t.Types.Select(type => (type, Activator.CreateInstance(t.solverType))))
-            .ToDictionary(t => Il2CppType.From(t.type).FullName, t => (MinigameSolver) t.Item2);
-    }
-
-    private static readonly Dictionary<string, MinigameSolver> MinigameSolvers;
-
-    public static bool CanComplete(Minigame minigame) => MinigameSolvers.ContainsKey(minigame.GetIl2CppType().FullName);
-    public static IEnumerator Complete(Minigame minigame, PlayerTask task)
-    {
-        if (!CanComplete(minigame)) throw new ArgumentException($"Cannot solve minigame {minigame.name}");
-
-        MinigameSolver solver = MinigameSolvers[minigame.GetIl2CppType().FullName];
-
-        InGameCursor.Instance.HideWhen(() => !minigame);
-
-        yield return new WaitForSeconds(0.4f);
-        yield return solver.CompleteMinigame(minigame, task);
-
-        // By this point we expect the solver to have completed the minigame,
-        // which means that it will close and this coroutine will not execute
-        // any code below.
-    }
 
     public abstract IEnumerator CompleteMinigame(Minigame minigame, PlayerTask task);
