+++ Neuro/Minigames/Solvers/DoorSabotagePolusSolver.cs
@@ -2,7 +2,7 @@
 using Neuro.Cursor;
 using UnityEngine;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(DoorBreakerGame))]
 public sealed class DoorSabotagePolusSolver : TasklessMinigameSolver<DoorBreakerGame>
