+++ Neuro/Minigames/Solvers/MeasureWeatherSolver.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using Neuro.Cursor;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(WeatherMinigame))]
 public sealed class MeasureWeatherSolver : MinigameSolver<WeatherMinigame>
