+++ Neuro/Minigames/Solvers/PrimeShieldsSolver.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using Neuro.Cursor;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(ShieldMinigame))]
 public sealed class PrimeShieldsSolver : MinigameSolver<ShieldMinigame>
