+++ Neuro/Minigames/Solvers/ProcessDataSolver.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using Neuro.Cursor;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(ProcessDataMinigame))]
 public sealed class ProcessDataSolver : MinigameSolver<ProcessDataMinigame>
