+++ Neuro/Minigames/Solvers/ReactorMeltdownSolver.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using Neuro.Cursor;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(ReactorMinigame))]
 public sealed class ReactorMeltdownSolver : TasklessMinigameSolver<ReactorMinigame>
