+++ Neuro/Minigames/Solvers/UnlockManifoldsSolver.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using Neuro.Cursor;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(UnlockManifoldsMinigame))]
 public sealed class UnlockManifoldsSolver : MinigameSolver<UnlockManifoldsMinigame>
