+++ Neuro/Minigames/Solvers/UploadDataSolver.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using Neuro.Cursor;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(UploadDataGame))]
 public sealed class UploadDataSolver : MinigameSolver<UploadDataGame>
