+++ Neuro/Minigames/Solvers/WaterPlantsSolver.cs
@@ -2,7 +2,7 @@
 using Neuro.Cursor;
 using UnityEngine;
 
-namespace Neuro.Minigames.Completion.Solvers;
+namespace Neuro.Minigames.Solvers;
 
 [MinigameSolver(typeof(WaterPlantsGame))]
 public sealed class WaterPlantsSolver : MinigameSolver<WaterPlantsGame>
